Update README.md
This commit adds a detailed README file to the repository, outlining the portfolio projects included, the tools and technologies used, and acknowledgments for Alex the Analyst’s tutorials, which inspired the projects. The README provides an overview of the purpose of the repository and guides users through the contents and projects.